Returning profiles list with sequelize orm.

+++ web/migrations/201607172054.temperature.log.js
@@ -12,17 +12,17 @@ module.exports = {
             },
 
             fahrenheit: {
-                type: DataTypes.INTEGER,
+                type: DataTypes.REAL,
                 allowNull: true
             },
 
             celsius: {
-                type: DataTypes.INTEGER,
+                type: DataTypes.REAL,
                 allowNull: true
             },
 
             kelvin: {
-                type: DataTypes.INTEGER,
+                type: DataTypes.REAL,
                 allowNull: true
             },
